+++ src/migration/migration.manager.ts
@@ -1,5 +1,8 @@
+import AlterTable from './types/altertable';
+import CreateTable from './types/createtable';
 import DbManager from '../dbs/dbmanager';
 import Migration from './migration';
+import MigrationType from './types/migrationtype';
 import fs from 'fs';
 import path from 'path';
 
@@ -14,6 +17,64 @@ export default class MigrationManager {
     this.config = config;
   }
 
+  public createMigration = async (migration: CreateTable | AlterTable) => {
+    const now = new Date();
+    const fileName = `[${now
+      .toISOString()
+      .replace(/T/g, '-')
+      .replace(/:/g, '-')}]${migration.getName()}.js`;
+    const nbFile = fs
+      .readdirSync(path.resolve(this.config.migrationPath))
+      .length.toString();
+    fs.writeFileSync(
+      path.resolve(this.config.migrationPath, fileName),
+      migration.generateMigrationFile(nbFile),
+    );
+    const migr = new Migration(`${migration.getName()}-${nbFile}`, 'up');
+    const model = DbManager.get().get();
+    if (!model) {
+      throw new Error('Database not found');
+    }
+    await migr.execute(model);
+  };
+
+  public analyzeMigrations = async (tableName: string) => {
+    const model = DbManager.get().get();
+    if (!model) {
+      throw new Error('Database not found');
+    }
+    // ! TODO change this
+    const migrations: string[] = (
+      await model.query(
+        'SELECT name FROM migration ORDER BY migrated_at ASC, id ASC',
+      )
+    ).map((m: any) => m.name);
+    const result: MigrationType[] = [];
+    const res = fs.readdirSync(this.config.migrationPath);
+    res.forEach((migrationFile: string) => {
+      const migrationPath = path.resolve(
+        this.config.migrationPath,
+        migrationFile,
+      );
+      const migrationRequired = require(migrationPath);
+      if (migrations.includes(migrationRequired.name)) {
+        const migration = new Migration(migrationRequired.name, 'up');
+        migrationRequired.up(migration);
+        const r = migration.findByTableName(tableName);
+        r.forEach((m) => {
+          result.push(m);
+        });
+      }
+    });
+    if (!result.length) return new CreateTable(tableName);
+    if (result[0].type !== 'createtable') return new CreateTable(tableName);
+    const globalMigration = result[0] as CreateTable;
+    for (let i = 1; i < result.length; i += 1) {
+      globalMigration.applyMigration(result[i]);
+    }
+    return globalMigration;
+  };
+
   /**
    * Make all migrations or given targetMigration
    * (call the up function of every migration)
@@ -24,6 +85,12 @@ export default class MigrationManager {
     if (!model) {
       throw new Error('Database not found');
     }
+    // ! TODO change this
+    const migrations: string[] = (
+      await model.query(
+        'SELECT name FROM migration ORDER BY migrated_at ASC, id ASC',
+      )
+    ).map((m: any) => m.name);
     const res = fs.readdirSync(this.config.migrationPath);
     res.sort();
     await res.reduce(async (prev, migrationFile) => {
@@ -33,7 +100,10 @@ export default class MigrationManager {
         migrationFile,
       );
       const migrationRequired = require(migrationPath);
-      if (!targetMigration || targetMigration === migrationRequired.name) {
+      if (
+        !migrations.includes(migrationRequired.name) &&
+        (!targetMigration || targetMigration === migrationRequired.name)
+      ) {
         const migration = new Migration(migrationRequired.name, 'up');
         console.log(`\x1b[35m## Migrating ${migrationRequired.name}\x1b[0m`);
         migrationRequired.up(migration);