+++ src/migration/types/altertable.ts
@@ -3,12 +3,10 @@ import { Field, FieldType } from './createtable';
 import MigrationType from './migrationtype';
 
 export default class AlterTable extends MigrationType {
-  private addedFields: Field[] = [];
-  private removedFields: string[] = [];
-  private tableName: string;
+  public addedFields: Field[] = [];
+  public removedFields: string[] = [];
   constructor(tableName: string) {
-    super();
-    this.tableName = tableName;
+    super(tableName, 'altertable');
   }
 
   public addField = (fieldName: string, fieldType: FieldType) => {
@@ -32,5 +30,23 @@ export default class AlterTable extends MigrationType {
     return {
       query: [...toRemove, ...toAdd],
     };
+  };
+
+  public generateMigrationFile = (name: string) => {
+    let file = `const Migration = require('@smallprod/models').Migration;\n\n /**\n *\n * @param {Migration} migration\n */\nconst up = (migration) => {\n    const alter = migration.alterTable('${this.tableName}');\n\n`;
+    this.removedFields.forEach((field) => {
+      file = `${file}   alter.removeField('${field}');\n`;
+    });
+    this.addedFields.forEach((field) => {
+      file = `${file}   alter${field.generateMigrationFile()};\n`;
+    });
+    file = `${file}};\n\n`;
+    file = `${file}/*\n * @param {Migration} migration\n */\nconst down = (migration) => {\n    migration.dropTable('${this.tableName}');\n};\n\n`;
+    file = `${file}module.exports = {\n   name: '${this.getName()}-${name}',\n    up,\n   down,\n};`;
+    return file;
+  };
+
+  public getName = () => {
+    return `alter-table-${this.tableName.toLowerCase()}`;
   };
 }